requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
pgvector==0.4.1
python-dotenv==1.1.1
sentence-transformers==5.1.0
pandas==2.3.2
